# Conflicts:
#	src/main/java/uk/gov/hmcts/opal/service/DraftAccountReferenceValidationService.java

+++ src/main/java/uk/gov/hmcts/opal/dto/legacy/LegacyDefendantAccountImpositionsResponseCommon.java
@@ -5,6 +5,7 @@ import jakarta.xml.bind.annotation.XmlAccessType;
 import jakarta.xml.bind.annotation.XmlAccessorType;
 import jakarta.xml.bind.annotation.XmlElement;
 import jakarta.xml.bind.annotation.XmlRootElement;
+import java.math.BigInteger;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -21,7 +22,7 @@ public class LegacyDefendantAccountImpositionsResponseCommon {
 
     @JsonProperty("version")
     @XmlElement(name = "version")
-    private Long version;
+    private BigInteger version;
 
     @JsonProperty("impositions")
     @XmlElement(name = "impositions")
